Ignore workspace members correctly
