+++ tests/fixtures/coverage-reports/real1/workspace_root.json
@@ -1,1 +1,1 @@
-{"data":[{"files":[{"filename":"src/lib.rs","summary":{"branches":{"count":0,"covered":0,"notcovered":0,"percent":0},"functions":{"count":4,"covered":4,"percent":100},"instantiations":{"count":4,"covered":4,"percent":100},"lines":{"count":13,"covered":11,"percent":84.615384615384613},"regions":{"count":9,"covered":7,"notcovered":2,"percent":77.777777777777786}}}],"totals":{"branches":{"count":0,"covered":0,"notcovered":0,"percent":0},"functions":{"count":4,"covered":4,"percent":100},"instantiations":{"count":4,"covered":4,"percent":100},"lines":{"count":13,"covered":11,"percent":84.615384615384613},"regions":{"count":9,"covered":7,"notcovered":2,"percent":77.777777777777786}}}],"type":"llvm.coverage.json.export","version":"2.0.1"}
+{"data":[{"files":[{"filename":"src/lib.rs","summary":{"branches":{"count":0,"covered":0,"notcovered":0,"percent":0},"functions":{"count":4,"covered":4,"percent":100},"instantiations":{"count":4,"covered":4,"percent":100},"lines":{"count":15,"covered":13,"percent":86.666666666666671},"regions":{"count":9,"covered":7,"notcovered":2,"percent":77.777777777777786}}}],"totals":{"branches":{"count":0,"covered":0,"notcovered":0,"percent":0},"functions":{"count":4,"covered":4,"percent":100},"instantiations":{"count":4,"covered":4,"percent":100},"lines":{"count":15,"covered":13,"percent":86.666666666666671},"regions":{"count":9,"covered":7,"notcovered":2,"percent":77.777777777777786}}}],"type":"llvm.coverage.json.export","version":"2.0.1"}
