+++ tests/fixtures/coverage-reports/real1/workspace_root.lcov.info
@@ -3,6 +3,6 @@ FNF:4
 FNH:4
 BRF:0
 BFH:0
-LF:13
-LH:11
+LF:15
+LH:13
 end_of_record
