+++ tests/fixtures/coverage-reports/real1/workspace_root.summary.txt
@@ -1,5 +1,5 @@
 Filename                      Regions    Missed Regions     Cover   Functions  Missed Functions  Executed       Lines      Missed Lines     Cover    Branches   Missed Branches     Cover
 -----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-src/lib.rs                          9                 2    77.78%           4                 0   100.00%          13                 2    84.62%           0                 0         -
+src/lib.rs                          9                 2    77.78%           4                 0   100.00%          15                 2    86.67%           0                 0         -
 -----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-TOTAL                               9                 2    77.78%           4                 0   100.00%          13                 2    84.62%           0                 0         -
+TOTAL                               9                 2    77.78%           4                 0   100.00%          15                 2    86.67%           0                 0         -
