+++ tests/fixtures/coverage-reports/real1/workspace_root.txt
@@ -12,4 +12,6 @@
    11|      1|fn test() {
    12|      1|    func(1);
    13|      1|    func(3);
-   14|      1|}
+   14|      1|    member1::func(0);
+   15|      1|    member2::func(0);
+   16|      1|}
