+++ tests/fixtures/coverage-reports/real1/workspace_root_all.json
@@ -1,0 +1,1 @@
+{"data":[{"files":[{"filename":"member1/member2/src/lib.rs","summary":{"branches":{"count":0,"covered":0,"notcovered":0,"percent":0},"functions":{"count":2,"covered":2,"percent":100},"instantiations":{"count":3,"covered":2,"percent":66.666666666666657},"lines":{"count":8,"covered":5,"percent":62.5},"regions":{"count":7,"covered":4,"notcovered":3,"percent":57.142857142857139}}},{"filename":"member1/src/lib.rs","summary":{"branches":{"count":0,"covered":0,"notcovered":0,"percent":0},"functions":{"count":2,"covered":2,"percent":100},"instantiations":{"count":3,"covered":2,"percent":66.666666666666657},"lines":{"count":8,"covered":5,"percent":62.5},"regions":{"count":7,"covered":4,"notcovered":3,"percent":57.142857142857139}}},{"filename":"src/lib.rs","summary":{"branches":{"count":0,"covered":0,"notcovered":0,"percent":0},"functions":{"count":4,"covered":4,"percent":100},"instantiations":{"count":4,"covered":4,"percent":100},"lines":{"count":15,"covered":13,"percent":86.666666666666671},"regions":{"count":9,"covered":7,"notcovered":2,"percent":77.777777777777786}}}],"totals":{"branches":{"count":0,"covered":0,"notcovered":0,"percent":0},"functions":{"count":8,"covered":8,"percent":100},"instantiations":{"count":10,"covered":8,"percent":80},"lines":{"count":31,"covered":23,"percent":74.193548387096769},"regions":{"count":23,"covered":15,"notcovered":8,"percent":65.217391304347828}}}],"type":"llvm.coverage.json.export","version":"2.0.1"}
