+++ tests/fixtures/coverage-reports/real1/workspace_root_all.lcov.info
@@ -1,0 +1,24 @@
+SF:member1/member2/src/lib.rs
+FNF:2
+FNH:2
+BRF:0
+BFH:0
+LF:8
+LH:5
+end_of_record
+SF:member1/src/lib.rs
+FNF:2
+FNH:2
+BRF:0
+BFH:0
+LF:8
+LH:5
+end_of_record
+SF:src/lib.rs
+FNF:4
+FNH:4
+BRF:0
+BFH:0
+LF:15
+LH:13
+end_of_record
