+++ tests/fixtures/coverage-reports/real1/workspace_root_all.summary.txt
@@ -1,0 +1,7 @@
+Filename                       Regions    Missed Regions     Cover   Functions  Missed Functions  Executed       Lines      Missed Lines     Cover    Branches   Missed Branches     Cover
+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+member1/member2/src/lib.rs           7                 3    57.14%           2                 0   100.00%           8                 3    62.50%           0                 0         -
+member1/src/lib.rs                   7                 3    57.14%           2                 0   100.00%           8                 3    62.50%           0                 0         -
+src/lib.rs                           9                 2    77.78%           4                 0   100.00%          15                 2    86.67%           0                 0         -
+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+TOTAL                               23                 8    65.22%           8                 0   100.00%          31                 8    74.19%           0                 0         -
