+++ tests/fixtures/coverage-reports/real1/workspace_root_all.txt
@@ -1,0 +1,64 @@
+member1/member2/src/lib.rs:
+    1|      1|pub fn func(x: u32) {pub fn func(x: u32) {
+    2|      1|    match x {
+    3|      1|        0 => {}
+    4|      0|        1 => {}
+    5|      0|        2 => {}
+    6|      0|        _ => {}
+    7|       |    }
+    8|      1|}
+  ------------------
+  | member2::func:
+  |    1|      1|pub fn func(x: u32) {
+  |    2|      1|    match x {
+  |    3|      1|        0 => {}
+  |    4|      0|        1 => {}
+  |    5|      0|        2 => {}
+  |    6|      0|        _ => {}
+  |    7|       |    }
+  |    8|      1|}
+  ------------------
+  | Unexecuted instantiation: member2::func
+  ------------------
+
+member1/src/lib.rs:
+    1|      1|pub fn func(x: u32) {pub fn func(x: u32) {
+    2|      1|    match x {
+    3|      1|        0 => {}
+    4|      0|        1 => {}
+    5|      0|        2 => {}
+    6|      0|        _ => {}
+    7|       |    }
+    8|      1|}
+  ------------------
+  | member1::func:
+  |    1|      1|pub fn func(x: u32) {
+  |    2|      1|    match x {
+  |    3|      1|        0 => {}
+  |    4|      0|        1 => {}
+  |    5|      0|        2 => {}
+  |    6|      0|        _ => {}
+  |    7|       |    }
+  |    8|      1|}
+  ------------------
+  | Unexecuted instantiation: member1::func
+  ------------------
+
+src/lib.rs:
+    1|      2|pub fn func(x: u32) {pub fn func(x: u32) {
+              ^1
+    2|      2|    match x {
+    3|      0|        0 => {}
+    4|      1|        1 => {}
+    5|      0|        2 => {}
+    6|      1|        _ => {}
+    7|       |    }
+    8|      2|}
+    9|       |
+   10|      1|#[test]
+   11|      1|fn test() {
+   12|      1|    func(1);
+   13|      1|    func(3);
+   14|      1|    member1::func(0);
+   15|      1|    member2::func(0);
+   16|      1|}
