+++ tests/fixtures/coverage-reports/real1/workspace_root_member2.json
@@ -1,0 +1,1 @@
+{"data":[{"files":[{"filename":"member1/member2/src/lib.rs","summary":{"branches":{"count":0,"covered":0,"notcovered":0,"percent":0},"functions":{"count":2,"covered":1,"percent":50},"instantiations":{"count":2,"covered":1,"percent":50},"lines":{"count":8,"covered":1,"percent":12.5},"regions":{"count":7,"covered":1,"notcovered":6,"percent":14.285714285714285}}}],"totals":{"branches":{"count":0,"covered":0,"notcovered":0,"percent":0},"functions":{"count":2,"covered":1,"percent":50},"instantiations":{"count":2,"covered":1,"percent":50},"lines":{"count":8,"covered":1,"percent":12.5},"regions":{"count":7,"covered":1,"notcovered":6,"percent":14.285714285714285}}}],"type":"llvm.coverage.json.export","version":"2.0.1"}
