+++ tests/fixtures/coverage-reports/real1/workspace_root_member2.lcov.info
@@ -1,0 +1,8 @@
+SF:member1/member2/src/lib.rs
+FNF:2
+FNH:1
+BRF:0
+BFH:0
+LF:8
+LH:1
+end_of_record
