+++ tests/fixtures/coverage-reports/real1/workspace_root_member2.summary.txt
@@ -1,0 +1,5 @@
+Filename                       Regions    Missed Regions     Cover   Functions  Missed Functions  Executed       Lines      Missed Lines     Cover    Branches   Missed Branches     Cover
+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+member1/member2/src/lib.rs           7                 6    14.29%           2                 1    50.00%           8                 7    12.50%           0                 0         -
+------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+TOTAL                                7                 6    14.29%           2                 1    50.00%           8                 7    12.50%           0                 0         -
