+++ tests/fixtures/coverage-reports/real1/workspace_root_member2.txt
@@ -1,0 +1,8 @@
+    1|      1|pub fn func(x: u32) {pub fn func(x: u32) {
+    2|      0|    match x {
+    3|      0|        0 => {}
+    4|      0|        1 => {}
+    5|      0|        2 => {}
+    6|      0|        _ => {}
+    7|       |    }
+    8|      0|}
