+++ tests/fixtures/coverage-reports/virtual1/exclude.json
@@ -1,1 +1,1 @@
-{"data":[{"files":[{"filename":"member1/src/lib.rs","summary":{"branches":{"count":0,"covered":0,"notcovered":0,"percent":0},"functions":{"count":4,"covered":4,"percent":100},"instantiations":{"count":4,"covered":4,"percent":100},"lines":{"count":14,"covered":12,"percent":85.714285714285708},"regions":{"count":9,"covered":7,"notcovered":2,"percent":77.777777777777786}}}],"totals":{"branches":{"count":0,"covered":0,"notcovered":0,"percent":0},"functions":{"count":4,"covered":4,"percent":100},"instantiations":{"count":4,"covered":4,"percent":100},"lines":{"count":14,"covered":12,"percent":85.714285714285708},"regions":{"count":9,"covered":7,"notcovered":2,"percent":77.777777777777786}}}],"type":"llvm.coverage.json.export","version":"2.0.1"}
+{"data":[{"files":[{"filename":"member1/src/lib.rs","summary":{"branches":{"count":0,"covered":0,"notcovered":0,"percent":0},"functions":{"count":4,"covered":4,"percent":100},"instantiations":{"count":4,"covered":4,"percent":100},"lines":{"count":16,"covered":14,"percent":87.5},"regions":{"count":9,"covered":7,"notcovered":2,"percent":77.777777777777786}}}],"totals":{"branches":{"count":0,"covered":0,"notcovered":0,"percent":0},"functions":{"count":4,"covered":4,"percent":100},"instantiations":{"count":4,"covered":4,"percent":100},"lines":{"count":16,"covered":14,"percent":87.5},"regions":{"count":9,"covered":7,"notcovered":2,"percent":77.777777777777786}}}],"type":"llvm.coverage.json.export","version":"2.0.1"}
