+++ tests/fixtures/coverage-reports/virtual1/exclude.lcov.info
@@ -3,6 +3,6 @@ FNF:4
 FNH:4
 BRF:0
 BFH:0
-LF:14
-LH:12
+LF:16
+LH:14
 end_of_record
