+++ tests/fixtures/coverage-reports/virtual1/exclude.summary.txt
@@ -1,5 +1,5 @@
 Filename                      Regions    Missed Regions     Cover   Functions  Missed Functions  Executed       Lines      Missed Lines     Cover    Branches   Missed Branches     Cover
 -----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-member1/src/lib.rs                  9                 2    77.78%           4                 0   100.00%          14                 2    85.71%           0                 0         -
+member1/src/lib.rs                  9                 2    77.78%           4                 0   100.00%          16                 2    87.50%           0                 0         -
 -----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-TOTAL                               9                 2    77.78%           4                 0   100.00%          14                 2    85.71%           0                 0         -
+TOTAL                               9                 2    77.78%           4                 0   100.00%          16                 2    87.50%           0                 0         -
