+++ tests/fixtures/coverage-reports/virtual1/workspace_root.json
@@ -1,1 +1,1 @@
-{"data":[{"files":[{"filename":"member1/src/lib.rs","summary":{"branches":{"count":0,"covered":0,"notcovered":0,"percent":0},"functions":{"count":4,"covered":4,"percent":100},"instantiations":{"count":4,"covered":4,"percent":100},"lines":{"count":14,"covered":12,"percent":85.714285714285708},"regions":{"count":9,"covered":7,"notcovered":2,"percent":77.777777777777786}}},{"filename":"member2/src/lib.rs","summary":{"branches":{"count":0,"covered":0,"notcovered":0,"percent":0},"functions":{"count":5,"covered":5,"percent":100},"instantiations":{"count":7,"covered":5,"percent":71.428571428571431},"lines":{"count":20,"covered":15,"percent":75},"regions":{"count":15,"covered":10,"notcovered":5,"percent":66.666666666666657}}}],"totals":{"branches":{"count":0,"covered":0,"notcovered":0,"percent":0},"functions":{"count":9,"covered":9,"percent":100},"instantiations":{"count":11,"covered":9,"percent":81.818181818181827},"lines":{"count":34,"covered":27,"percent":79.411764705882348},"regions":{"count":24,"covered":17,"notcovered":7,"percent":70.833333333333343}}}],"type":"llvm.coverage.json.export","version":"2.0.1"}
+{"data":[{"files":[{"filename":"member1/src/lib.rs","summary":{"branches":{"count":0,"covered":0,"notcovered":0,"percent":0},"functions":{"count":4,"covered":4,"percent":100},"instantiations":{"count":4,"covered":4,"percent":100},"lines":{"count":16,"covered":14,"percent":87.5},"regions":{"count":9,"covered":7,"notcovered":2,"percent":77.777777777777786}}},{"filename":"member2/member3/src/lib.rs","summary":{"branches":{"count":0,"covered":0,"notcovered":0,"percent":0},"functions":{"count":2,"covered":2,"percent":100},"instantiations":{"count":3,"covered":2,"percent":66.666666666666657},"lines":{"count":8,"covered":5,"percent":62.5},"regions":{"count":7,"covered":4,"notcovered":3,"percent":57.142857142857139}}},{"filename":"member2/src/lib.rs","summary":{"branches":{"count":0,"covered":0,"notcovered":0,"percent":0},"functions":{"count":5,"covered":5,"percent":100},"instantiations":{"count":7,"covered":5,"percent":71.428571428571431},"lines":{"count":20,"covered":15,"percent":75},"regions":{"count":15,"covered":10,"notcovered":5,"percent":66.666666666666657}}},{"filename":"member2/src/member4/src/lib.rs","summary":{"branches":{"count":0,"covered":0,"notcovered":0,"percent":0},"functions":{"count":2,"covered":2,"percent":100},"instantiations":{"count":3,"covered":2,"percent":66.666666666666657},"lines":{"count":8,"covered":5,"percent":62.5},"regions":{"count":7,"covered":4,"notcovered":3,"percent":57.142857142857139}}}],"totals":{"branches":{"count":0,"covered":0,"notcovered":0,"percent":0},"functions":{"count":13,"covered":13,"percent":100},"instantiations":{"count":17,"covered":13,"percent":76.470588235294116},"lines":{"count":52,"covered":39,"percent":75},"regions":{"count":38,"covered":25,"notcovered":13,"percent":65.789473684210535}}}],"type":"llvm.coverage.json.export","version":"2.0.1"}
