+++ tests/fixtures/coverage-reports/virtual1/workspace_root.lcov.info
@@ -3,8 +3,16 @@ FNF:4
 FNH:4
 BRF:0
 BFH:0
-LF:14
-LH:12
+LF:16
+LH:14
+end_of_record
+SF:member2/member3/src/lib.rs
+FNF:2
+FNH:2
+BRF:0
+BFH:0
+LF:8
+LH:5
 end_of_record
 SF:member2/src/lib.rs
 FNF:5
@@ -13,4 +21,12 @@ BRF:0
 BFH:0
 LF:20
 LH:15
+end_of_record
+SF:member2/src/member4/src/lib.rs
+FNF:2
+FNH:2
+BRF:0
+BFH:0
+LF:8
+LH:5
 end_of_record
