+++ tests/fixtures/coverage-reports/virtual1/workspace_root.summary.txt
@@ -1,6 +1,8 @@
-Filename                      Regions    Missed Regions     Cover   Functions  Missed Functions  Executed       Lines      Missed Lines     Cover    Branches   Missed Branches     Cover
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-member1/src/lib.rs                  9                 2    77.78%           4                 0   100.00%          14                 2    85.71%           0                 0         -
-member2/src/lib.rs                 15                 5    66.67%           5                 0   100.00%          20                 5    75.00%           0                 0         -
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-TOTAL                              24                 7    70.83%           9                 0   100.00%          34                 7    79.41%           0                 0         -
+Filename                           Regions    Missed Regions     Cover   Functions  Missed Functions  Executed       Lines      Missed Lines     Cover    Branches   Missed Branches     Cover
+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+member1/src/lib.rs                       9                 2    77.78%           4                 0   100.00%          16                 2    87.50%           0                 0         -
+member2/member3/src/lib.rs               7                 3    57.14%           2                 0   100.00%           8                 3    62.50%           0                 0         -
+member2/src/lib.rs                      15                 5    66.67%           5                 0   100.00%          20                 5    75.00%           0                 0         -
+member2/src/member4/src/lib.rs           7                 3    57.14%           2                 0   100.00%           8                 3    62.50%           0                 0         -
+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+TOTAL                                   38                13    65.79%          13                 0   100.00%          52                13    75.00%           0                 0         -
